requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-docx
pymupdf
sentence-transformers
langchain==0.3.27
langchain-chroma==0.2.4
langchain-community==0.3.25
langchain-core==0.3.74
langchain-groq==0.3.7
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.9
chromadb
python-dotenv